Fix use dash variable names in command line option
when wsrep_sst_method is set to mariabackup, the donor node makes a backup with mariabackup, then streams to the joining node. Variables with underscores are invalid as command line options, cause mariabackup to fail on the donor node and disconnect from the joining node. This is the culprit of nodes failing to join the cluster after a crash. The log of the joining node only indicates this is a network connection issue. When mariabackup is successful with revised valid options, failed nodes can receive streaming of write sets and can join the cluster successfully.